+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+* readme.txt
+* fix PHP error if no plugins with `Requires Plugins` header found
 
 #### 0.11.6.4
 * plugin to date with new changelog